+++ packages/alfa-graph/src/graph.ts
@@ -9,12 +9,12 @@ import { Set } from "@siteimprove/alfa-set";
 /**
  * @public
  */
-export class Graph<T>
+export class Graph<T, O extends Serializable.Options = Serializable.Options>
   implements
     Iterable<[T, Iterable<T>]>,
     Equatable,
     Hashable,
-    Serializable<Graph.JSON<T>>
+    Serializable<Graph.JSON<T>, O>
 {
   public static of<T>(nodes: Map<T, Set<T>>): Graph<T> {
     return new Graph(nodes);
@@ -214,10 +214,10 @@ export class Graph<T>
     return [...this].map(([node, neighbors]) => [node, [...neighbors]]);
   }
 
-  public toJSON(): Graph.JSON<T> {
+  public toJSON(options?: O): Graph.JSON<T> {
     return this.toArray().map(([node, neighbors]) => [
-      Serializable.toJSON(node),
-      neighbors.map((node) => Serializable.toJSON(node)),
+      Serializable.toJSON(node, options),
+      neighbors.map((node) => Serializable.toJSON(node, options)),
     ]);
   }