+++ packages/alfa-sequence/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @siteimprove/alfa-sequence
 
+## 0.82.0
+
+### Minor Changes
+
+- **Added:** Serialization options are now accepted, and passed on, by `toJSON()` on these types. ([#1622](https://github.com/Siteimprove/alfa/pull/1622))
+
+- **Breaking:** Node 18 is no longer supported. ([#1618](https://github.com/Siteimprove/alfa/pull/1618))
+
 ## 0.81.0
 
 ### Patch Changes